apkcrawler: Click on elements.
Change the device connection setup so that we can click on the first
clickable element and then recursively call the function. Because of
this, we end up in some loop behaviors that will be fixed in the future.

Store the screenshot with a unique filename in the format
/apkcrawler/<package name>/<activity name>##.png

With the change in device connection, we also can no longer use the
AndroidViewClient screenshot capabilities.

Change-Id: Icd7e7796e8a74b31f7e7b8e700a47c04512cc73f