apkcrawler: Create View class.
View class contains all necessary components of a view hierarchy: the
activity/fragment names, view hierarchy, list of clickable components,
and a pointer to the screenshot file.

Right now we store the current view in a View variable, but do not
keep track of all of the views. The next step will be store all of
the views in a data structure and check for duplicates.

Change-Id: Iad5e723f8feabcecb767dc30ca663b7f24cab27c