capsule: Use AndroidViewClient commands, fix logic.
AndroidViewClient uses hard-coded ADB paths, which aren't always
included in the output of 'which -a adb' so use the AVC path. Also,
change many of the commands to be AVC commands.

Also, don't save the view if we are no longer in the app, and attempt
to get back to the app if we leave.

Change-Id: I9156e0181759c133f55e5e60912d33a8d753b9d6