luma/rico: Add post-processing scripts.
Add two scripts. One is a utility for debugging the view hierarchies
produced by rico. It queries an attached device and gets the view
hierarchy and draws the clickable elements on it. The other script is
for post-processing all the data produced by Rico and combines them
to produce visualizations for the entire user trace.

Change-Id: Ia8b11186f6dd8802414c68f3d943c0572eeeaa63