capsule: Return to unexhausted Views.
For all Views that are not completely explored after our initial crawl,
we attempt to return to those Views and click on the rest of the
components. Once a View has no more clickable components or if we
cannot follow the path back to the View, we remove it from the queue.

Each time we restart the crawl, we kill and then restart the app to
return to the root view and minimize the amount of changed state data
in the app.

I also made some other tweaks to fix stability.

This fixes #2.

Change-Id: I7bf4f528a3825b01b66e181600bca7374483664c