Capsule: bug fixes and recrawl command line option.
Fixed errors that crashed Capsule during specific apps-- specifically
that follow_path_to_view won't be looped, prevent crashing when there
is no view hierarchy, and throw an exception when we get unicode
characters for the x/y position of components.

Also, allow the user to specify if an app should be recrawled if there's
already data for it.

Change-Id: Iaa400d20953e8a637ce7b448cf77248e28eca174