capsule: Allow Facebook logins and crawl in-order.
Automate the Facebook authorization login flow. When a Facebook login
component (or a clickable element with the words facebook and login) is
found, it prioritizes this element and clicks on it, then clicks where
the Continue button should be on the authorization WebView.

Also, crawl the components in order instead of reverse order because
sometimes drawer elements are considered part of a hierarchy and the
first element often opens the drawer, so this increases our chance of
clicking on elements in the drawer. If the first component
contains the word 'back,' click on it last so that we explore the View
before going back.

Change-Id: Ic3cdf53db022b825e52031bb6786b189d295e304