capsule: Allow G+ logins, improve fb logins.
Automate the Google+ login flow and allow extra clicks on the Facebook
authorization WebView in case the app wants to post to the user's wall.

Also, improve the resiliency of the click function. For some reason, the
AndroidViewClient touch function does not always register clicks since
it relies on using the res id.

Change-Id: I54f87b5f453b1631624456d2fd7f2a112640fa13