capsule: Allow multiple connected devices.
Currently, the adb calls in Capsule fail if more than one device is
connected because adb command needs to specify the serial number of the
intended device. Instead, we pass the serial number of the device that
Capsule is running on to crawlpkg so that the commands are issued to the
proper device.

Change-Id: Ied65930bd319868589e28f2a208a667d20101be6