capsule: Set layout depth.
For each layout, set the depth as distance from the root Layout.
Also, a few fixes to following the path to return to a Layout, to
improve following a path if we leave the app, and prevent loops where we
don't get to intended Layout but end on one with the same activity name
and fragment list.

Change-Id: I7d19e0dc7f0ccc6b002c73b9de7358fd88d2078f