luma: CrowdSTF Manager and HIT Service
This manager adds an adminstration layer on top of CrowdSTF to generate
tokens for crowd workers to use apps on remote devices via HITs
on Amazon Turk. It includes the ability to poll devices continuously
and create HITs based on their usage and availablity. A usage stream
visualizes HIT events in real time.

Change-Id: Ice1fa45d60b45b2c4fe542346fdd0721382a183c