capsule: Add config file and type text in fields.
Introduce a Config class that reads user info from a config.ini file and
uses the field names to help populate text fields.

During our crawl, whenever we see a text field, try to assess what the
field is looking for by its resource id. If we find a match with a key
in the config file, type the value of that field. Else, type the default
text.

Change-Id: Ifb6821eddd2c085aea1066183a11158ea63094bb