capsule: Add settings to config, more reserved words.
Allow the user to decide if capsule should lock the phone to portrait
mode and if it should clear notifications (to make the screenshots a
little nicer) and split up the config file into sections.  Also, add a
group of words associated with already being registered for an app to
the list so that Views with these words in the text or id are clicked on
last.

Change-Id: Ic33d656510ed2bca05f618aea5e041df6c0aa852